requirements.txt
httpx[http2]
flask
ddddocr
numpy
Pillow
onnxruntime
cryptography
